Move Coordinator worker startup code
The move from constructor to a method leaves open the possibility
of providing a message based start for a job rather than having
it happen as soon as a Coordinator is created. I will probably
make that switch and in the meantime it makes everything a little
cleaner in the constructor.