Add Job as top-level actor that a user interacts with (#4)
Coordinator is now _Coordinator and no longer exposed to the fork/join
end-user. Having "Job" be the primary interface that users interact
with will make writing documentation easier as explaining what the
Coordinator did from an end-user perspective was a little odd as it
did "many things". It still does a few different things but that
will be further broken down.

This is an initial first step, I'm going to move all the creation
of actors that the Coordinator currently does into the Job.